backend/requirements.txt
pydantic>=2.0
pandas>=2.0
fastapi>=0.110
uvicorn>=0.29
python-multipart>=0.0.9
pymongo>=4.8
google-generativeai>=0.3.0
reportlab>=4.0.0
